+++ aprsd/utils.py
@@ -15,7 +15,7 @@ DEFAULT_CONFIG_DICT = {
     "aprs": {
         "login": "someusername",
         "password": "somepassword",
-        "host": "noam.aprs2.net",
+        "host": "rotate.aprs.net",
         "port": 14580,
         "logfile": "/tmp/arsd.log",
     },
@@ -152,8 +152,8 @@ def parse_config(config_file):
     )
     check_option(config, "aprs", "login")
     check_option(config, "aprs", "password")
-    check_option(config, "aprs", "host")
-    check_option(config, "aprs", "port")
+    # check_option(config, "aprs", "host")
+    # check_option(config, "aprs", "port")
     check_option(config, "aprs", "logfile", "./aprsd.log")
     check_option(config, "imap", "host")
     check_option(config, "imap", "login")
