+++ dev-requirements.in
@@ -7,4 +7,3 @@ pep8-naming
 black
 isort
 Sphinx
-thesmuggler
