+++ dev-requirements.txt
@@ -10,7 +10,7 @@ attrs==20.3.0             # via pytest
 babel==2.9.0              # via sphinx
 black==20.8b1             # via -r dev-requirements.in
 certifi==2020.12.5        # via requests
-chardet==3.0.4            # via requests
+chardet==4.0.0            # via requests
 click==7.1.2              # via black
 coverage==5.3             # via pytest-cov
 distlib==0.3.1            # via virtualenv
@@ -40,7 +40,7 @@ pytest-cov==2.10.1        # via -r dev-requirements.in
 pytest==6.2.0             # via -r dev-requirements.in, pytest-cov
 pytz==2020.4              # via babel
 regex==2020.11.13         # via black
-requests==2.25.0          # via sphinx
+requests>=2.25.0          # via sphinx
 six==1.15.0               # via tox, virtualenv
 snowballstemmer==2.0.0    # via sphinx
 sphinx==3.3.1             # via -r dev-requirements.in
