+++ requirements.in
@@ -7,3 +7,4 @@ pyyaml
 six
 requests
 thesmuggler
+aprslib
