--- test-requirements-py2.txt
@@ -1,3 +1,0 @@
-flake8
-pytest
-mock
