--- test-requirements.txt
@@ -1,2 +1,0 @@
-flake8
-pytest
