+++ tests/test_main.py
@@ -4,7 +4,7 @@ import unittest
 
 import pytest
 
-from aprsd import main
+from aprsd import email
 
 if sys.version_info >= (3, 2):
     from unittest import mock
@@ -13,11 +13,11 @@ else:
 
 
 class testMain(unittest.TestCase):
-    @mock.patch("aprsd.main._imap_connect")
-    @mock.patch("aprsd.main._smtp_connect")
+    @mock.patch("aprsd.email._imap_connect")
+    @mock.patch("aprsd.email._smtp_connect")
     def test_validate_email(self, imap_mock, smtp_mock):
         """Test to make sure we fail."""
         imap_mock.return_value = None
         smtp_mock.return_value = {"smaiof": "fire"}
 
-        main.validate_email()
+        email.validate_email()
