+++ tox.ini
@@ -2,7 +2,7 @@
 minversion = 2.9.0
 skipdist = True
 skip_missing_interpreters = true
-envlist = py{27,36,37,38},pep8,fmt-check
+envlist = pep8,py{36,37,38},fmt-check
 
 # Activate isolated build environment. tox will use a virtual environment
 # to build a source distribution from the source tree. For build tools and
@@ -14,7 +14,6 @@ setenv = VIRTUAL_ENV={envdir}
 usedevelop = True
 install_command = pip install {opts} {packages}
 deps = -r{toxinidir}/requirements.txt
-       -r{toxinidir}/test-requirements.txt
        -r{toxinidir}/dev-requirements.txt
 commands =
     # Use -bb to enable BytesWarnings as error to catch str/bytes misuse.
@@ -23,28 +22,9 @@ commands =
 #        --cov="{envsitepackagesdir}/aprsd" --cov-report=html --cov-report=term {posargs}
     {envpython} -bb -m pytest {posargs}
 
-[testenv:py27]
-setenv = VIRTUAL_ENV={envdir}
-usedevelop = True
-install_command = pip install {opts} {packages}
-deps = -r{toxinidir}/requirements.txt
-       -r{toxinidir}/test-requirements-py2.txt
-commands =
-    # Use -bb to enable BytesWarnings as error to catch str/bytes misuse.
-    # Use -Werror to treat warnings as errors.
-#    {envpython} -bb -Werror -m pytest \
-#        --cov="{envsitepackagesdir}/aprsd" --cov-report=html --cov-report=term {posargs}
-    {envpython} -bb -Werror -m pytest
-
 [testenv:docs]
 deps = -r{toxinidir}/test-requirements.txt
 commands = sphinx-build -b html docs/source docs/html
-
-[testenv:pep8-27]
-deps = -r{toxinidir}/requirements.txt
-       -r{toxinidir}/test-requirements-py2.txt
-commands =
-  flake8 {posargs} aprsd
 
 [testenv:pep8]
 commands =
@@ -78,7 +58,6 @@ exclude = .venv,.git,.tox,dist,doc,.ropeproject
 # This section is not needed if not using GitHub Actions for CI.
 [gh-actions]
 python =
-    2.7: py27, pep8-27
     3.6: py36, pep8, fmt-check
     3.7: py38, pep8, fmt-check
     3.8: py38, pep8, fmt-check, type-check, docs
